+++ src/Render/draws/index.ts
@@ -1,2 +1,4 @@
 export * from "./background";
 export * from "./ruler";
+export * from "./scaleInfo";
+export * from "./referenceLine";