+++ src/Render/interactions/index.ts
@@ -1,2 +1,3 @@
 export { enableDrag } from "./drag";
 export { enableZoom } from "./zoom";
+export { enableReferenceLine } from "./referenceLine";